Validate StageProp, SequenceChannel and PlaylistSequence UUIDs on server (#28)

+++ persistence/src/main/java/io/sparkled/persistence/playlist/impl/query/SavePlaylistSequencesQuery.java
@@ -3,6 +3,7 @@ package io.sparkled.persistence.playlist.impl.query;
 import io.sparkled.model.entity.Playlist;
 import io.sparkled.model.entity.PlaylistSequence;
 import io.sparkled.model.validator.PlaylistSequenceValidator;
+import io.sparkled.model.validator.exception.EntityValidationException;
 import io.sparkled.persistence.PersistenceQuery;
 import io.sparkled.persistence.QueryFactory;
 import org.slf4j.Logger;
@@ -29,14 +30,32 @@ public class SavePlaylistSequencesQuery implements PersistenceQuery<Void> {
     @Override
     public Void perform(QueryFactory queryFactory) {
         PlaylistSequenceValidator playlistSequenceValidator = new PlaylistSequenceValidator();
+        playlistSequences.forEach(ps -> ps.setPlaylistId(playlist.getId()));
         playlistSequences.forEach(playlistSequenceValidator::validate);
 
-        EntityManager entityManager = queryFactory.getEntityManager();
-        playlistSequences.forEach(entityManager::merge);
-        logger.info("Saved {} playlist sequence(s) for playlist {}.", playlistSequences.size(), playlist.getId());
+        if (uuidAlreadyInUse(queryFactory)) {
+            throw new EntityValidationException("Playlist sequence already exists on another playlist.");
+        } else {
+            EntityManager entityManager = queryFactory.getEntityManager();
+            playlistSequences.forEach(entityManager::merge);
+            logger.info("Saved {} playlist sequence(s) for playlist {}.", playlistSequences.size(), playlist.getId());
 
-        deleteRemovedPlaylistSequences(queryFactory);
-        return null;
+            deleteRemovedPlaylistSequences(queryFactory);
+            return null;
+        }
+    }
+
+    private boolean uuidAlreadyInUse(QueryFactory queryFactory) {
+        List<UUID> uuidsToCheck = playlistSequences.stream().map(PlaylistSequence::getUuid).collect(toList());
+        uuidsToCheck = uuidsToCheck.isEmpty() ? noUuids : uuidsToCheck;
+
+        long uuidsInUse = queryFactory.select(qPlaylistSequence)
+                .from(qPlaylistSequence)
+                .where(
+                        qPlaylistSequence.playlistId.ne(playlist.getId()).and(qPlaylistSequence.uuid.in(uuidsToCheck))
+                )
+                .fetchCount();
+        return uuidsInUse > 0;
     }
 
     private void deleteRemovedPlaylistSequences(QueryFactory queryFactory) {
