+++ persistence/src/main/java/io/sparkled/persistence/sequence/impl/query/SaveSequenceChannelsQuery.java
@@ -1,8 +1,10 @@
 package io.sparkled.persistence.sequence.impl.query;
 
+import com.querydsl.core.BooleanBuilder;
 import io.sparkled.model.entity.Sequence;
 import io.sparkled.model.entity.SequenceChannel;
 import io.sparkled.model.validator.SequenceChannelValidator;
+import io.sparkled.model.validator.exception.EntityValidationException;
 import io.sparkled.persistence.PersistenceQuery;
 import io.sparkled.persistence.QueryFactory;
 import org.slf4j.Logger;
@@ -29,14 +31,30 @@ public class SaveSequenceChannelsQuery implements PersistenceQuery<Void> {
     @Override
     public Void perform(QueryFactory queryFactory) {
         SequenceChannelValidator sequenceChannelValidator = new SequenceChannelValidator();
+        sequenceChannels.forEach(sc -> sc.setSequenceId(sequence.getId()));
         sequenceChannels.forEach(sequenceChannelValidator::validate);
 
-        final EntityManager entityManager = queryFactory.getEntityManager();
-        sequenceChannels.forEach(entityManager::merge);
-        logger.info("Saved {} sequence channel(s) for sequence {}.", sequenceChannels.size(), sequence.getId());
+        if (uuidAlreadyInUse(queryFactory)) {
+            throw new EntityValidationException("Sequence channel already exists on another sequence.");
+        } else {
+            final EntityManager entityManager = queryFactory.getEntityManager();
+            sequenceChannels.forEach(entityManager::merge);
+            logger.info("Saved {} sequence channel(s) for sequence {}.", sequenceChannels.size(), sequence.getId());
 
-        deleteRemovedSequenceChannels(queryFactory);
-        return null;
+            deleteRemovedSequenceChannels(queryFactory);
+            return null;
+        }
+    }
+
+    private boolean uuidAlreadyInUse(QueryFactory queryFactory) {
+        List<UUID> uuidsToCheck = sequenceChannels.stream().map(SequenceChannel::getUuid).collect(toList());
+        uuidsToCheck = uuidsToCheck.isEmpty() ? noUuids : uuidsToCheck;
+
+        long uuidsInUse = queryFactory.select(qSequenceChannel)
+                .from(qSequenceChannel)
+                .where(qSequenceChannel.sequenceId.ne(sequence.getId()).and(qSequenceChannel.uuid.in(uuidsToCheck)))
+                .fetchCount();
+        return uuidsInUse > 0;
     }
 
     private void deleteRemovedSequenceChannels(QueryFactory queryFactory) {
